feat: utils functions for testing.
